pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-team-spec"
version = "0.1.0"
description = "Multi-agent AI system spec — Orchestrator, Planner, Builder, Critic, Platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "openhands-ai>=1.5.0",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"
